+++ superset/databases/schemas.py
@@ -246,7 +246,11 @@ class DatabaseParametersSchemaMixin:
         the constructed SQLAlchemy URI to be passed.
         """
         parameters = data.pop("parameters", None)
-        encrypted_extra = data.get("encrypted_extra", None)
+        serialized_encrypted_extra = data.get("encrypted_extra", "{}")
+        try:
+            encrypted_extra = json.loads(serialized_encrypted_extra)
+        except json.decoder.JSONDecodeError:
+            encrypted_extra = {}
 
         if parameters:
             if "engine" not in parameters:
@@ -277,9 +281,11 @@ class DatabaseParametersSchemaMixin:
                     ]
                 )
 
-            data["sqlalchemy_uri"] = engine_spec.build_sqlalchemy_uri(
+        if hasattr(engine_spec, "build_sqlalchemy_uri"):
+            data["sqlalchemy_uri"] = engine_spec.build_sqlalchemy_uri(  # type: ignore
                 parameters, encrypted_extra
             )
+
         return data
 
 
@@ -556,3 +562,15 @@ class ImportV1DatabaseSchema(Schema):
         password = make_url(uri).password
         if password == PASSWORD_MASK and data.get("password") is None:
             raise ValidationError("Must provide a password for the database")
+
+
+class EncryptedField(fields.String):
+    pass
+
+
+def encrypted_field_properties(self, field: Any, **_) -> Dict[str, Any]:  # type: ignore
+    ret = {}
+    if isinstance(field, EncryptedField):
+        if self.openapi_version.major > 2:
+            ret["x-encrypted-extra"] = True
+    return ret
